Revert "Revert motd, clear, and user action response"
This reverts commit 3144bd7cf8ac48d5339ef07942e932ad7d4233de.

The reason this commit had to be made was because certain apis were PR'd
upstream, but were unwanted and thus removed in a separate commit. It
would likely have been better to drop the commits pushed upstream but
it's too late to fix that, so now we have a revert of a revert.